updates ResultMessage interface with an index signature (#1216)
By including an index signature, the ResultMessage becomes more extensible.
Plugins are allowed to push messages that have properties besides the few
that are already defined. In fact, the only ones that are now specified
explicitly, are those that are required. This includes `plugin`, which
previously was not marked required. However, the API documentation guides
users to always specify both `type` and `plugin`.

see: http://api.postcss.org/Result.html#messages